+++ CHANGELOG.md
@@ -18,6 +18,7 @@ x.x.x Release notes (yyyy-MM-dd)
 * Upgraded Realm Core from v12.0.0 to v12.1.0.
 * Fix for updated FLX sync error message. ([#4611](https://github.com/realm/realm-js/pull/4611))
 * Updated build script to use Xcode 13.1 to match latest Apple App Store compatibility. ([#4605](https://github.com/realm/realm-js/issues/4605))
+* Removed support for node.js v12 and earlier versions and bumped to Node-API v5.
 
 10.18.0 Release notes (2022-5-29)
 =============================================================